Using UNITCHECK and DATA section
The default mapping data is now stored in the DATA section of the module
instead of a perl hash. It is easier to update from outside sources if
required.

The UNITCHECK phase hook is used to parse the DATA section into a perl
hash and make it available for the constructor